refactor: coupon 비즈니스 로직 model 파일로 분리
